se agrega contratos en su propio archivo - se agrega loading
