+++ _components/master/comments/contracts/comments.ts
@@ -1,0 +1,11 @@
+export interface CommentModelContract {
+    text: string | null;
+    textEdit: string;
+    active: boolean;
+    user: string;
+    loading: boolean;
+    avatar: string;
+}
+export interface EditorConfigContract {
+    height: number;
+}
